Initial import of malli.dev.virhe as virhe.core
This version is a direct import of malli.dev.virhe namespace included
in the release 0.14 of Malli. It can be used as its own independent
library to implement dev mode error messages similar to those that are
used in the new 0.14 release of Malli. The repository includes tests
and test runners for both Clojure and ClojureScript builds.